fixed, energy.py needs work
