+++ source_files/ScanFunctions_IEX.py
@@ -2,10 +2,9 @@ from math import *
 import time
 import datetime
 import subprocess
-import argparse
+
 import re
-import csv
-import ast
+
 import socket
 import numpy as np
 import numpy.polynomial.polynomial as poly
@@ -284,85 +283,7 @@ def M3R_Table(branch,motor):   # WARNING: branch_pv uses: D => (Tx <= 0) and (Ry
 ############# WARNING: when changing table, make sure the string sequence below is still relevant:
 
     
-def centroid(t=None,q=1): 
-    '''
-    Return position of centroid, sigma, m3r:RY (mirror pitch)
-    Optional argument t to set camera intergration time.
-    '''
-    if t is not None:
-        caput('29id_ps6:cam1:AcquireTime',t)
-    else:
-        t=caget('29id_ps6:cam1:AcquireTime')
-    position=  round(caget('29id_ps6:Stats1:CentroidX_RBV'),2)
-    sigma=round(caget('29id_ps6:Stats1:SigmaX_RBV'),2)
-    intensity=round(caget('29id_ps6:Stats1:CentroidTotal_RBV'),2)
-    m3rRY=round(caget("29id_m3r:RY_MON"),4)
-    if q != None:
-        print('(position, sigma, total intensity, integration time (s), mirror pitch):')
-    return position,sigma,intensity,t,m3rRY
 
-
-
-def align_m3r(p=118,debug=False):
-    
-    def mvm3r_pitch_FR(x):
-        if -16.53<x<-16.38:
-            motor_pv="29id_m3r:RY_POS_SP"
-            #print('... moving to '+str(x))
-            caput(motor_pv,x)
-            sleep(0.5)
-            caput("29id_m3r:MOVE_CMD.PROC",1)
-            sleep(1)
-            #print('Positioned')
-        else:
-            print('Out of range')
-
-    shutter=caget('PA:29ID:SDS_BLOCKING_BEAM.VAL')
-    camera=caget('29id_ps6:cam1:Acquire',as_string=True)
-    hv=caget('29idmono:ENERGY_SP')
-    if shutter == 'ON':
-        foo=input_d('Shutter D is closed, do you want to open it (Y or N)? >')
-        if foo == 'Y'.lower() or foo == 'y' or foo == 'yes'.lower():
-            Open_DShutter()
-        else:
-            print('Aborting...')
-            return
-    if camera == 'Done':
-        caput('29id_ps6:cam1:Acquire','Acquire') 
-     
-    if hv < 2190:
-        caput('29id_ps6:cam1:AcquireTime',0.001)
-        caput('29id_ps6:Stats1:CentroidThreshold',7)
-    else:
-        caput('29id_ps6:cam1:AcquireTime',0.03)
-        caput('29id_ps6:Stats1:CentroidThreshold',7)
-    sleep(1)    
-    intensity=caget('29id_ps6:Stats1:CentroidTotal_RBV')
-    if intensity < 10000:
-        print('Count too low. Please adjust camera settings')
-    else:   
-        print('Starting...')
-        mvm3r_pitch_FR(-16.52)
-        position = centroid(q=None)[0]
-        max_pitch=-16.39
-        #print('position = '+str(position))
-        while position < p:
-            position = centroid(q=None)[0]
-            if position < p-10:
-                pitch = caget('29id_m3r:RY_POS_SP')
-                mvm3r_pitch_FR(min(pitch + 0.005,max_pitch))
-                #print(centroid(q=None)[4])
-                position = centroid(q=None)[0]
-                #print('position = '+str(position))
-            elif  p-10<= position:
-                pitch = caget('29id_m3r:RY_POS_SP')
-                mvm3r_pitch_FR(min(pitch + 0.001,max_pitch))
-                #print(centroid(q=None)[4])
-                position = centroid(q=None)[0]
-                #print('position = '+str(position))
-        print('Done')
-        print(centroid())
-        print('\n')
 
     
 def GoToSlit1AScribe():
@@ -442,24 +363,11 @@ def WireScan(which,scanIOC=None,diag='In',**kwargs):
 
 """
 
-def read_dict(FileName,FilePath="/home/beams22/29IDUSER/Documents/User_Macros/Macros_29id/IEX_Dictionaries/"):
-    with open(join(FilePath, FileName)) as f:
-        for c,line in enumerate(f.readlines()):
-            if line[0] == '=':
-                lastdate=line[8:16]
-            lastline=line
-        mydict=ast.literal_eval(lastline)
-    return mydict
 
 
 
-def ID_Coef(grt,mode,hv):    # Mode = state (0=RCP,1=LCP,2=V,3=H); 
-    
-    """Return the ID coeff for a given polarization mode and energy;
-    with Mode = 0 (RCP),1 (LCP), 2 (V), 3 (H).
-    Current coefficient dictionary:
-        /home/beams22/29IDUSER/Documents/User_Macros/Macros_29id/IEX_Dictionaries/Dict_IDCal.txt
-    """
+
+
 
     
     def ListRange(grt,mode,IDdict):  # extract the list of break pts for a given mode/grt 
@@ -497,38 +405,10 @@ def ID_Coef(grt,mode,hv):    # Mode = state (0=RCP,1=LCP,2=V,3=H);
 
 
 
-def ID_Calc(grt,mode,hv):    # Mode = state (0=RCP,1=LCP,2=V,3=H)
-    """Calculate the ID SP for a given polarization mode and energy;
-    with Mode = 0 (RCP),1 (LCP), 2 (V), 3 (H)"""
-    if type(mode)== str:
-        mode={'RCP':0,'LCP':1,'V':2,'H':3}[mode]
-    try:
-        K=ID_Coef(grt,mode,hv)
-        #ID = K[0] + K[1]*hv**1 + K[2]*hv**2 + K[3]*hv**3 + K[4]*hv**4 + K[5]*hv**5
-        ID=poly.polyval(hv,K)
-    except KeyError:
-        print("WARNING: PLease select one of the following:")
-        print("        mode 0 = RCP")
-        print("        mode 1 = LCP")
-        print("        mode 2 = V")
-        print("        mode 3 = H")
-        ID=caget("ID29:EnergySeteV")
-    return round(ID,1)
 
 
 
 
-def ID_Range():      # mode = state (0=RCP,1=LCP,2=V,3=H)
-    Mode=caget("ID29:ActualMode")
-    GRT=caget("29idmono:GRT_DENSITY")
-    ID_min=[400,400,440,250,250]
-    ID_max=3800
-    hv_min=[390,390,430,245,245]
-    if GRT == 2400:
-        hv_max=2000
-    elif GRT == 1200:
-        hv_max=3000
-    return ID_min[Mode],hv_min[Mode],hv_max,ID_max
 
 
 def Energy_Range(grating,IDmode):
@@ -538,33 +418,6 @@ def Energy_Range(grating,IDmode):
     energy_min, energy_max = BL_range[grating][IDmode]
     return energy_min, energy_max
 
-
-
-
-
-def Slit_Coef(n):
-    if n == 1:
-        pv='29id:k_slit1A'
-        #Redshifted x (H):
-        H0=2.3325
-        H1=-.000936
-        H2=2.4e-7
-         #Redshifted z (V):
-        V0=2.3935
-        V1=-.0013442
-        V2=3.18e-7
-    if n == 2:
-        pv='29id:k_slit2B'
-        #Redshifted x (H):
-        H0=3.61
-        H1=-0.00186
-        H2=5.2e-7
-        #Redshifted z (V):
-        V0=6.8075
-        V1=-0.003929
-        V2=9.5e-7
-    K=H0,H1,H2,V0,V1,V2
-    return pv,K
 
 
 
@@ -607,19 +460,7 @@ def Monitor_Beam(which='user'):
     print('Sending text/email')
 
 
-def WaitForPermission():
-    """
-    Monitors the ID permissions and waits for the ID to be in User Mode and then breaks
-    Checks the status every 30 seconds
-    """
-    while True:
-        ID_Access=caget("ID29:AccessSecurity.VAL")
-        if (ID_Access!=0):
-            print("Checking ID permission, please wait..."+dateandtime())
-            sleep(30)
-        else:
-            print("ID now in user mode -"+dateandtime())
-            break
+
 
 def WaitForBeam():
     """
@@ -636,67 +477,7 @@ def WaitForBeam():
             break
 
 
-def ID_Ready(q=None):
-    while True:
-        checkready=caget("ID29:feedback.VAL")
-        checkbusy=caget("ID29:BusyRecord")
-        if (checkready!="Ready")|(checkbusy==1):
-            sleep(2)
-        else:
-            break
-    if q is None:
-        print("ID Ready")
 
-def ID_Stop():
-    WaitForPermission()
-    caput("ID29:Main_on_off.VAL",1,wait=True,timeout=18000)
-
-
-def ID_Start(Mode):
-    "Starts ID with a specific polarization; Mode = \"H\", \"V\", \"RCP\" or \"LCP\""
-    WaitForPermission()
-    print("Starting ID  -  "+dateandtime())
-    caput("ID29:EnergySet.VAL",3.8)
-    caput("ID29:Main_on_off.VAL",0,wait=True,timeout=18000)
-    ID_Ready()
-    Switch_IDMode(Mode)
-    while True:
-        SS1=caget("EPS:29:ID:SS1:POSITION")
-        SS2=caget("EPS:29:ID:SS2:POSITION")
-        PS2=caget("EPS:29:ID:PS2:POSITION")
-        check=SS1*SS2*PS2
-        if (check!=8):
-            print("MAIN SHUTTER CLOSED !!!" , dateandtime())
-            Open_MainShutter()
-            sleep(60)
-        else:
-            print("Shutter is now open" , dateandtime())
-            break
-
-
-def ID_Restart():
-    Mode=caget("ID29:ActualMode")
-    print("Restarting ID", dateandtime())
-    caput("ID29:Main_on_off.VAL",1,wait=True,timeout=18000)
-    sleep(10)
-    while True:
-        RBV=caget("ID29:Energy.VAL")
-        checkready=caget("ID29:feedback.VAL")
-        checkbusy=caget("ID29:BusyRecord")
-        if (checkready!="Ready") or (RBV < 3.7):
-            sleep(2)
-        elif ((checkready=="Ready") and (RBV > 3.7)) and (checkbusy=="Busy"):
-            caput("ID29:Busy.VAL",0)
-        else:
-            break
-    sleep(10)
-    caput("ID29:EnergySet.VAL",3.8)
-    caput("ID29:Main_on_off.VAL",0,wait=True,timeout=18000)
-    ID_Ready()
-    caput("ID29:DesiredMode.VAL",Mode,wait=True,timeout=18000)
-    sleep(10)
-
-    print("ID is back ON", dateandtime())
 
 
 
@@ -712,29 +493,6 @@ def Check_IDMode(which):
             print("Switching ID mode, please wait...")
             caput("ID29:DesiredMode.VAL",Mode,wait=True,timeout=18000)     # RCP
             ID_Ready()
-    else:
-        print("WARNING: Not a valid polarization mode, please select one of the following:")
-        print(" \"RCP\", \"LCP\", \"H\", \"V\" ")
-
-
-def Switch_IDMode(which):
-    "Change ID polarization; which = \"H\", \"V\", \"RCP\" or \"LCP\""
-    GRT=caget("29idmono:GRT_DENSITY")
-    if which in ["RCP","LCP","V","H"]:
-        Mode=ID_State2Mode("Mode",which)
-        Check_MainShutter()
-        ActualMode=caget("ID29:ActualMode")
-        if Mode != ActualMode:
-            print("Turning ID off...")
-            ID_Stop()
-            
-            sleep(10)
-            caput("ID29:EnergySet.VAL",3.8)
-            caput("ID29:Main_on_off.VAL",0,wait=True,timeout=18000)
-            print("Switching ID mode, please wait...")
-            caput("ID29:DesiredMode.VAL",Mode,wait=True,timeout=18000)     # RCP
-            ID_Ready()
-        print("ID Mode:",which)
     else:
         print("WARNING: Not a valid polarization mode, please select one of the following:")
         print(" \"RCP\", \"LCP\", \"H\", \"V\" ")
@@ -759,84 +517,6 @@ def Switch_IDQP(energy,mode,QP):
 
 ##### ID energy :
 
-def SetID(hv):
-    "Sets optimum ID set point for hv(eV) (max intensity)"
-    Check_MainShutter()
-    GRT=Check_Grating()
-    hv_SP=round(SetRange(hv),3)
-    ID_min=ID_Range()[0]
-    if hv_SP == round(hv,2) and hv_SP < 3001:
-        Mode=caget("ID29:ActualMode")
-        ID_SP=max(ID_Calc(GRT,Mode,hv_SP),ID_min)
-        ID_SP_RBV=round(caget("ID29:EnergySet.VAL"),4)*1000
-        if ID_SP == ID_SP_RBV:                    # compare ID (eV)SP to the (keV)SP energy
-            ID_RBV=caget("ID29:EnergyRBV")
-            print("\nID SET : "+"%.1f" % ID_SP, "eV")
-            print("ID RBV : "+"%.1f" % ID_RBV, "eV")
-            print(caget('ID29:TableDirection',as_string=True))
-        else:
-            caput("ID29:EnergyScanSeteV",ID_SP,wait=True,timeout=18000)
-            sleep(1)
-            caput("ID29:EnergyScanSeteV",ID_SP+0.001,wait=True,timeout=18000)
-
-            ID_Ready()
-            print("\nID SET : "+"%.1f" % ID_SP, "eV")
-            sleep(5)
-            ID_RBV=caget("ID29:EnergyRBV")
-            print("ID RBV : "+"%.1f" % ID_RBV, "eV")
-            ID_DIF=abs(ID_RBV-ID_SP)
-            ID_BW=ID_SP*0.095
-            if ID_DIF>ID_BW:
-                sleep(20)
-                ID_RBV=caget("ID29:EnergyRBV")
-                print("ID RBV : "+"%.1f" % ID_RBV, "eV")
-                ID_DIF=abs(ID_RBV-ID_SP)
-                if ID_DIF>ID_BW:
-                    ID_Restart()
-                    SetID_keV_pV(hv_SP)
-            print(caget('ID29:TableDirection',as_string=True))
-#    elif  2450 < hv_SP < 3000:
-#        print("Not calibrated above 2450 eV. ID set point needs to be determined manually.")
-    else:
-        print("Please select a different energy.")
-
-def SetID_keV_pV(hv_eV):
-    "Sets optimum ID set point for hv(eV) (max intensity) using KeV PVs"
-    Check_MainShutter()
-    hv_SP=SetRange(hv_eV)                # check hv range
-    GRT=Check_Grating()
-    if hv_SP == hv_eV:
-        Mode=caget("ID29:ActualMode")
-        ID_SP=ID_Calc(GRT,Mode,hv_SP)/1000.0        # ID_SP in keV
-        ID_SP_RBV=round(caget("ID29:EnergySet.VAL"),3)    # ID_SP_RBV in keV
-        if ID_SP == ID_SP_RBV:                # checks if ID is already at the SP energy
-            ID_RBV=caget("ID29:Energy.VAL")
-            print("\nID SET : "+"%.3f" % ID_SP, "keV")
-            print("ID RBV : "+"%.3f" % ID_RBV, "keV")
-            print(caget('ID29:TableDirection',as_string=True))
-        else:
-            caput("ID29:EnergyScanSet",ID_SP,wait=True,timeout=18000)
-            sleep(1)
-            caput("ID29:EnergyScanSet",ID_SP+0.0001,wait=True,timeout=18000)
-        
-            ID_Ready()
-            print("\nID SET : "+"%.3f" % ID_SP, "keV")
-            sleep(5)
-            ID_RBV=caget("ID29:Energy.VAL")
-            print("ID RBV : "+"%.3f" % ID_RBV, "keV")
-            ID_DIF=abs(ID_RBV-ID_SP)
-            ID_BW=ID_SP*0.095
-            if ID_DIF > ID_BW:
-                sleep(20)
-                ID_RBV=caget("ID29:Energy.VAL")
-                print("\nID RBV : "+"%.3f" % ID_RBV, "keV")
-                ID_DIF=abs(ID_RBV-ID_SP)
-                if ID_DIF > ID_BW:
-                    ID_Restart()
-                    SetID_keV_pV(hv_eV)
-            print(caget('ID29:TableDirection',as_string=True))
-    else:
-        print("Please select a different energy.")
 
 
 def SetID_Raw(hv):
@@ -910,24 +590,6 @@ def Get_MainShutter():
 
 
 
-def Check_MainShutter():
-    "Checks main shutter is open, if not opens it"
-    while True:
-        SS1=caget("EPS:29:ID:SS1:POSITION")
-        SS2=caget("EPS:29:ID:SS2:POSITION")
-        PS2=caget("EPS:29:ID:PS2:POSITION")
-        check=SS1*SS2*PS2
-        if (check!=8):
-            print("MAIN SHUTTER CLOSED !!!" , dateandtime())
-            WaitForPermission()
-            Open_MainShutter()
-            sleep(10)
-        else:
-            break
-    ID_off=caget("ID29:Main_on_off.VAL")
-    if ID_off == 1:
-        ID_Start("RCP")
-        sleep(30)
 
 
 
@@ -967,15 +629,7 @@ def Close_BranchShutter():
     caput("PC:29ID:S"+branch+"S_CLOSE_REQUEST.VAL",1,wait=True,timeout=18000)
     print("Closing "+branch+"-Shutter...")
 
-def Close_CShutter():
-    branch="C"
-    caput("PC:29ID:S"+branch+"S_CLOSE_REQUEST.VAL",1,wait=True,timeout=18000)
-    print("Closing "+branch+"-Shutter...")
 
-def Close_DShutter():
-    branch="D"
-    caput("PC:29ID:S"+branch+"S_CLOSE_REQUEST.VAL",1,wait=True,timeout=18000)
-    print("Closing "+branch+"-Shutter...")
 
 def Close_CValve():
     branch="C"
@@ -1055,9 +709,6 @@ def Close_RSoXS():
             print('V10D valve now closed -- REMEMBER TO CLOSE V12D BY UNPLUGGING!!!!')
             break
 
-def Open_MainShutter():
-    caput("PC:29ID:FES_OPEN_REQUEST.VAL",1, wait=True,timeout=180000)
-    print("Opening Main Shutter...")
 
 def Open_BranchShutter():
     "Opens current branch shutter (based on deflecting mirror position)"
@@ -1074,10 +725,7 @@ def Open_CShutter():
     caput("PC:29ID:S"+branch+"S_OPEN_REQUEST.VAL",1,wait=True,timeout=18000)
     print("Opening "+branch+"-Shutter...")
 
-def Open_DShutter():
-    branch="D"
-    caput("PC:29ID:S"+branch+"S_OPEN_REQUEST.VAL",1,wait=True,timeout=18000)
-    print("Opening "+branch+"-Shutter...")
+
 
 def Open_CValve():
     branch="C"
@@ -1187,101 +835,6 @@ def Tweak_M3R_Pitch(sign):
         Tweak_M3R(Motor_Name,val)
     elif sign == "-":
         Tweak_M3R(Motor_Name,-val)
-
-def Switch_Branch(which, forced=False, noshutter=False,noreset=False,nocam=False):
-    """Switch beam into which = \"c\" or \"d\" branch (by retracting/inserting deflecting mirror)
-        Optionnal keyword argument:
-            forced = True      (default=False) => forces change branch even if the beamline think it is already there
-            noshutter = True   (default=False) => does not open/close shutters
-            noreset   = True   (default=False) => does not reset stuff
-            nocam = True   (default=False) => does not turn on/off cameras
-    """
-    which=which.upper()
-    if which in ["C","D","E"]:
-
-    # Check current branch:
-        Current_Branch=CheckBranch().upper()
-        if Current_Branch == which and forced:
-                Tx=round(caget("29id_m3r:TX_MON"),2)
-                Ty=round(caget("29id_m3r:TY_MON"),2)
-                Tz=round(caget("29id_m3r:TZ_MON"),2)
-                Rx=round(caget("29id_m3r:RX_MON"),2)
-                Ry=round(caget("29id_m3r:RY_MON"),2)
-                Rz=round(caget("29id_m3r:RZ_MON"),2)
-                print("    M3R @ "+"%.2f" % Tx, "/","%.2f" % Ty, "/","%.2f" % Tz, "/","%.2f" % Rx, "/","%.2f" % Ry, "/","%.2f" % Rz)
-                if Tx == 0 and Ty+Tz+Rx+Ry+Rz == 0:
-                    print("\nMirror homed...")
-                    print("...Try again using: Switch_Branch(which, forced=True)")
-                else:
-                    print("\nWell, looks like the beam is already in this branch...")
-                    print("...if you want to force the mirror motion, use the argument: forced=True")
-        else:
-
-        # Close both shutters:
-            print("\n")
-            print("Switching branch, please wait...")
-            if not noshutter:
-                Close_DShutter()
-                Close_CShutter()
-        # Move M3R:
-            motors=["TX","TY","TZ","RX","RY","RZ"]
-            for m in motors:
-                motor_pv="29id_m3r:"+m+"_POS_SP"
-                caput(motor_pv,M3R_Table(which,m))
-            sleep(2)
-            caput("29id_m3r:MOVE_CMD.PROC",1)
-            sleep(2)
-            while caget("29id_m3r:SYSTEM_STS")==0:
-                sleep(.5)
-        # Relax bellows by doing large Z translation:
-            Move_M3R("TY",5)
-            Move_M3R("TY",0)
-            #if which in ["D"]:
-                #Move_M3R("TX",-1)        # FR added on 2/2/2017 after adjusting GRT roll
-        # Print current positions:
-            print(dateandtime()," -  In "+which+"-Branch")
-            sleep(2)
-            Tx=caget("29id_m3r:TX_MON")
-            Ty=caget("29id_m3r:TY_MON")
-            Tz=caget("29id_m3r:TZ_MON")
-            Rx=caget("29id_m3r:RX_MON")
-            Ry=caget("29id_m3r:RY_MON")
-            Rz=caget("29id_m3r:RZ_MON")
-            print("\nM3R @ "+"%.2f" % Tx, "/","%.2f" % Ty, "/","%.2f" % Tz, "/","%.2f" % Rx, "/","%.2f" % Ry, "/","%.2f" % Rz)
-        # Open branch shutters:
-            if not noshutter:
-                shutter="PC:29ID:S"+which+"S_OPEN_REQUEST.VAL"
-                caput(shutter,1)
-                sleep(10)
-                print("Opening "+which+"-Shutter...")
-                Check_BranchShutter()
-            if not noreset:
-                scanIOC=M3R_Table(which,'scanIOC')
-                Reset_Scan(scanIOC)
-                Current_Branch=CheckBranch_Name()
-            if not nocam:
-                pvcam1=PV("29id_ps1:cam1:Acquire")
-                pvcam2=PV("29id_ps2:cam1:Acquire")
-                pvcam3=PV("29id_ps3:cam1:Acquire")
-                pvcam4=PV("29id_ps4:cam1:Acquire")
-                pvcam6=PV("29id_ps6:cam1:Acquire")
-                pvcam7=PV("29id_ps7:cam1:Acquire")
-                cam_list=[pvcam1,pvcam2,pvcam3,pvcam4,pvcam6,pvcam7]
-                cam_dict={'C':[0,1,2],'D':[3,4,6]}  # index of cam_list 
-                for pvcam in cam_list:                  # turn OFF all cam
-                    try:
-                        if pvcam.connected:  pvcam.put(0)
-                    except:
-                        pass
-                for i in cam_dict[which]:                 # turn ON relevant cam
-                    try:
-                        if cam_list[i].connected: cam_list[i].put(1)
-                        else: print(cam_list[i].pvname+' not connected')
-                    except:
-                        pass
-    else:
-        print("\nWARNING: Not a valid branch name, please select one of the following:")
-        print(" \"c\" for ARPES, \"d\" for Kappa ")
 
 
 
@@ -1428,76 +981,12 @@ def MeshD(In_Out):
 ###### Energy & Gratings:
 
 
-def SetRange(hv):
-    Mode=caget("ID29:ActualMode")
-    hv_min=ID_Range()[1]
-    hv_max=ID_Range()[2]
-    sleep(0.1)
-    hv_SP=min(round(max(hv,hv_min)*1.0,3),hv_max)*1.0
-    sleep(0.1)
-#    if round(hv,1) != hv_SP:
-    if hv < hv_min or hv > hv_max:
-        print("\nWARNING: Set point out of BL energy range:")
-    return hv_SP
-
-
-def SetBL(hv,c=1):
-    caput("29id:BeamlineEnergyAllow",'Disable',wait=True,timeout=18000)
-    caput("29id:BeamlineEnergySet",hv,wait=True,timeout=18000)
-    hv_SP=round(SetRange(hv),3)
-    if hv_SP == round(hv,3):
-        #if (hv>2250):
-            #ActualMode=caget("ID29:ActualMode")
-            #if(ActualMode<2):         #circular
-                #print("Setting ID to second harmonic.")
-        SetID(hv)
-        SetMono(hv)
-        SetSlit_BL(c2B=c,q='q')
-        print("\n")
-#        caput("29id:BeamlineEnergyAllow",'Enable',wait=True,timeout=18000)
-    else:
-        print("Please select a different energy.")
-
-
-def SetMono(eV):
-    GRT=Check_Grating()
-    if GRT == "HEG":
-        min_limit=120
-        max_limit=2200
-    elif GRT == "MEG":
-        min_limit=120
-        max_limit=3000
-    eV = max(min_limit,eV)
-    eV = min(max_limit,eV)
-    caput("29idmono:ENERGY_SP",eV,wait=True,timeout=60)
-    sleep(2.5)
-    STS=Mono_Status()
-    if STS >1:
-            print("Mono error - resetting motor")
-            caput("29idmonoMIR:P.STOP",1)
-            caput("29idmonoGRT:P.STOP",1)
-            sleep(1)
-            caput("29idmono:ENERGY_SP",eV,wait=True,timeout=18000)
-            sleep(2.5)
-            caput("29idmono:ENERGY_SP",eV,wait=True,timeout=18000)
-    print("Mono set to",str(eV),"eV")
 
 
 
-def Mono_Status():
-    MIR_Status=caget("29idmonoMIR:P_AXIS_STS")
-    GRT_Status=caget("29idmonoGRT:P_AXIS_STS")
-    Status=MIR_Status*GRT_Status
-    return Status
 
 
-def Check_Grating():
-    GRTd=caget("29idmono:GRT_DENSITY")
-    if GRTd == 1200:
-        GRT = "MEG"
-    elif GRTd == 2400:
-        GRT = "HEG"
-    return GRT
+
 
 
 def Move_GRT(which):
@@ -1644,11 +1133,7 @@ def Scan_Kappa_Go(scanIOC='Kappa',scanDIM=1,**kwargs):
     """
     Scan_Go(scanIOC,scanDIM=scanDIM,**kwargs)
     
-def Scan_ARPES_Go(scanIOC='ARPES',scanDIM=1,**kwargs):
-    """Starts the N dimension scan in the ARPES chamber (N=ScanDIM)
-    Logging is automatic: use **kwargs or the optional logging arguments see scanlog() for details
-    """
-    Scan_Go(scanIOC,scanDIM=scanDIM,**kwargs) 
+
 
 def Scan_RSoXS_Go(scanIOC='RSoXS',scanDIM=1,**kwargs):
     """Starts the N dimension scan in the RSoXS chamber using the Kappa IOC (N=ScanDIM)
@@ -3256,26 +2741,6 @@ def SetSlit(n,Hsize=None,Vsize=None,Hcenter=0,Vcenter=0,q=None):
     else:
         print('Not a valid slit number')
     
-def SetSlit1A(Hsize,Vsize,Hcenter,Vcenter,q=None):
-    caput("29idb:Slit1Hsync.PROC",1)
-    caput("29idb:Slit1Vsync.PROC",1)
-    caput("29idb:Slit1Hsize.VAL", Hsize)
-    caput("29idb:Slit1Vsize.VAL", Vsize, wait=True,timeout=18000)
-    caput("29idb:Slit1Hcenter.VAL",Hcenter)
-    caput("29idb:Slit1Vcenter.VAL",Vcenter, wait=True,timeout=18000)
-    if not q:
-        print("Slit-1A = ("+str(round(Hsize,3))+"x"+str(round(Vsize,3))+") @ ("+str(Hcenter)+","+str(Vcenter)+")")
-
-def SetSlit2B(Hsize,Vsize,Hcenter,Vcenter,q=None):
-    caput("29idb:Slit2Hsync.PROC",1)
-    caput("29idb:Slit2Vsync.PROC",1)
-    caput("29idb:Slit2Hsize.VAL", Hsize)
-    caput("29idb:Slit2Vsize.VAL", Vsize, wait=True,timeout=18000)
-    caput("29idb:Slit2Hcenter.VAL",Hcenter)
-    caput("29idb:Slit2Vcenter.VAL",Vcenter, wait=True,timeout=18000)
-    if not q:
-        print("Slit-2B = ("+str(Hsize)+"x"+str(Vsize)+") @ ("+str(Hcenter)+","+str(Vcenter)+")")
-
 
 
 def SetSlit2B_Small(Hsize,Vsize,Hcenter,Vcenter,coef):
@@ -3294,12 +2759,6 @@ def SetSlit2B_Small(Hsize,Vsize,Hcenter,Vcenter,coef):
 
 
 # Slits Fits:
-
-def Aperture_Fit(hv,n):
-    K=Slit_Coef(n)[1]
-    sizeH=K[0]+K[1]*hv+K[2]*hv*hv
-    sizeV=K[3]+K[4]*hv+K[5]*hv*hv
-    return [round(sizeH,3),round(sizeV,3)]
 
 
 def Slit3C_Fit(size):
@@ -3368,37 +2827,6 @@ def Scan_D5D(start=-62.5,stop=-42.5,step=0.5,**kwargs):
 
 
 
-
-def SetSlit_BL(c2B=1,c1A=1,q=None):
-
-    RBV=caget("29idmono:ENERGY_MON")
-    GRT=caget("29idmono:GRT_DENSITY")
-    hv=max(RBV,500)
-    hv=min(RBV,2000)
-    c=4.2/2.2
-
-    if GRT==1200:
-        GRT='MEG'
-        V=0.65        #  set to 65% of RR calculation for both grt => cf 2016_2_summary
-    elif GRT==2400:
-        GRT='HEG'
-        V=0.65*c        #  set to 65% of RR calculation (no longer 80%) => cf 2016_2_summary
-
-    try:
-        slit_position=read_dict(FileName='Dict_Slit.txt')
-    except KeyError:
-        print("Unable to read dictionary")
-        return
-
-    V2center= slit_position[GRT]['S2V']
-    H2center= slit_position[GRT]['S2H']
-    V1center= slit_position[GRT]['S1V']
-    H1center= slit_position[GRT]['S1H']
-
-    Size1A=( Aperture_Fit(hv,1)[0]*c1A,       Aperture_Fit(hv,1)[1]*c1A )
-    Size2B=( Aperture_Fit(hv,2)[0]*c2B, round(Aperture_Fit(hv,2)[1]*c2B*V,3))
-    SetSlit1A(Size1A[0],Size1A[1],H1center,V1center,q)    # standard operating
-    SetSlit2B(Size2B[0],Size2B[1],H2center,V2center,q)
 
 
 
@@ -6627,25 +6055,6 @@ def AbortScript():
 ##############################               BL commands            ##############################
 ##############################################################################################################
 
-def energy(val,c=1,m3r=True):
-    """
-    Sets the beamline energy: insertion device + mono + apertures.
-    Use c < 1 to kill flux density.
-    """
-    SetBL(val,c)
-    SetMono(val)
-    if m3r == True:
-        if CheckBranch() == 'd':
-            print('\nalign_m3r()')
-            try:
-                align_m3r()
-                sleep(1)
-                if caget('29id_m3r:RY_POS_SP') == -16.52:
-                    align_m3r(debug=True)
-            except:
-                print('Unable to align; check camera settings.')
-        
-
 
 def mono(val):
     """
@@ -6753,11 +6162,6 @@ def mvy(val):
     name="y"
     Move_Motor_vs_Branch(name,val)
 
-def mvz(val):
-    """ Moves z motor in the endstation specified by CheckBranch()
-    """
-    name="z"
-    Move_Motor_vs_Branch(name,val)
 
 def mvth(val):
     """ Moves phi motor in the endstation specified by CheckBranch()
@@ -6817,19 +6221,6 @@ def mvkphi(val):
 
         
         
-def uan(tth,th):
-    """ Moves tth and th motors simultaneously in the in the Kappa chamber
-    """
-    caput(Kappa_PVmotor("tth")[1],tth)
-    caput(Kappa_PVmotor("th")[1],th)
-    while True:
-        MotorBusy=caget("29idKappa:alldone.VAL",as_string=True)
-        if MotorBusy!='done':
-    #        print "No beam current, please wait..."+dateandtime()
-            sleep(0.1)
-        else:
-            print('tth='+str(tth)+' th='+str(th))
-            break
 
 def mvDetV(val):
     """ Moves the vetector vertically motor in the in the RSoXS chamber
@@ -7069,74 +6460,8 @@ def scanx(start,stop,step,mode="absolute",scanIOC=None,scanDIM=1,**kwargs):
     #elif mybranch == "e":
     #    Scan_RSoXS_Motor_Go("x",start,stop,step,mode=mode,scanIOC=scanIOC,scanDIM=scanDIM,**kwargs)        
 
-def scany(start,stop,step,mode="absolute",scanIOC=None,scanDIM=1,**kwargs):
-    mybranch=CheckBranch()
-    if scanIOC is None:
-        scanIOC=BL_ioc()
-    if mybranch == "c":
-        Scan_ARPES_Motor_Go("y",start,stop,step,mode=mode,**kwargs)
-    elif mybranch == "d":
-        Scan_Kappa_Motor_Go("y",start,stop,step,mode=mode,**kwargs)
-    #elif mybranch == "e":
-    #    Scan_RSoXS_Motor_Go("y",start,stop,step,mode=mode,scanIOC=scanIOC,scanDIM=scanDIM,**kwargs)
-
-def scanz(start,stop,step,mode="absolute",scanIOC=None,scanDIM=1,**kwargs):
-    mybranch=CheckBranch()
-    if scanIOC is None:
-        scanIOC=BL_ioc()
-    if mybranch == "c":
-        Scan_ARPES_Motor_Go("z",start,stop,step,mode=mode,**kwargs)
-    elif mybranch == "d":
-        Scan_Kappa_Motor_Go("z",start,stop,step,mode=mode,scanIOC=scanIOC,scanDIM=scanDIM,**kwargs)        
-    #elif mybranch == "e":
-    #    Scan_RSoXS_Motor_Go("z",start,stop,step,mode=mode,scanIOC=scanIOC,scanDIM=scanDIM,**kwargs)        
-
-def scanth(start,stop,step,mode="absolute",scanIOC=None,scanDIM=1,**kwargs):
-    mybranch=CheckBranch()
-    if scanIOC is None:
-        scanIOC=BL_ioc()
-    if mybranch == "c":
-        Scan_ARPES_Motor_Go("th",start,stop,step,mode=mode,**kwargs)        
-    elif mybranch == "d":
-        print("Scanning pseudo motor 29idKappa:Euler_Theta")
-        Scan_Kappa_Motor_Go("th",start,stop,step,mode=mode,scanIOC=scanIOC,scanDIM=scanDIM,**kwargs) 
-    #elif mybranch == "e":
-    #    Scan_RSoXS_Motor_Go("th",start,stop,step,mode=mode,scanIOC=scanIOC,scanDIM=scanDIM,**kwargs)        
 
 
-def scanchi(start,stop,step,mode="absolute",scanIOC=None,scanDIM=1,**kwargs):
-    mybranch=CheckBranch()
-    if scanIOC is None:
-        scanIOC=BL_ioc()
-    if mybranch == "c":
-        print("chi motor scan not implemented")
-    elif mybranch == "d":
-        Scan_Kappa_Motor_Go("chi",start,stop,step,mode=mode,scanIOC=scanIOC,scanDIM=scanDIM,**kwargs)        
-    elif mybranch =="e":
-        print("chi motor scan not implemented")
-
-def scanphi(start,stop,step,mode="absolute",scanIOC=None,scanDIM=1,**kwargs):
-    mybranch=CheckBranch()
-    if scanIOC is None:
-        scanIOC=BL_ioc()
-    if mybranch == "c":
-        print("phi motor scan not implemented")
-    elif mybranch == "d":
-        Scan_Kappa_Motor_Go("phi",start,stop,step,mode=mode,scanIOC=scanIOC,scanDIM=scanDIM,**kwargs)        
-    elif mybranch =="e":
-        print("phi motor scan not implemented")
-
-        
-def scantth(start,stop,step,mode="absolute",scanIOC=None,scanDIM=1,**kwargs):
-    mybranch=CheckBranch()
-    if scanIOC is None:
-        scanIOC=BL_ioc()
-    if mybranch == "c":
-        print("tth motor does not exit")
-    elif mybranch == "d":
-        Scan_Kappa_Motor_Go("tth",start,stop,step,mode=mode,scanIOC=scanIOC,scanDIM=scanDIM,**kwargs)        
-    #elif mybranch =="e":
-    #    Scan_RSoXS_Motor_Go("tth",start,stop,step,mode=mode,scanIOC=scanIOC,scanDIM=scanDIM,**kwargs)     
 
         
 def scankth(start,stop,step,mode="absolute",scanIOC=None,scanDIM=1,**kwargs):
@@ -7192,9 +6517,7 @@ def dscany(start,stop,step,scanIOC=None,scanDIM=1,**kwargs):
     '''RELATIVE scany'''
     scany(start,stop,step,"relative",scanIOC,scanDIM,**kwargs)
 
-def dscanz(start,stop,step,scanIOC=None,scanDIM=1,**kwargs):
-    '''RELATIVE scanz'''
-    scanz(start,stop,step,"relative",scanIOC,scanDIM,**kwargs)
+
 
 def dscantth(start,stop,step,scanIOC=None,scanDIM=1,**kwargs):
     '''RELATIVE scantth'''
@@ -7204,13 +6527,7 @@ def dscanth(start,stop,step,scanIOC=None,scanDIM=1,**kwargs):
     '''RELATIVE scanth'''
     scanth(start,stop,step,"relative",scanIOC,scanDIM,**kwargs)
 
-def dscanchi(start,stop,step,scanIOC=None,scanDIM=1,**kwargs):
-    '''RELATIVE scanchi'''
-    scanchi(start,stop,step,"relative",scanIOC,scanDIM,**kwargs)
 
-def dscanphi(start,stop,step,scanIOC=None,scanDIM=1,**kwargs):
-    '''RELATIVE scanphi'''
-    scanphi(start,stop,step,"relative",scanIOC,scanDIM,**kwargs)
 
 def dscankth(start,stop,step,scanIOC=None,scanDIM=1,**kwargs):
     '''RELATIVE scankth'''
